* Split in namespaces * read body once.
